+++ cmd/assistantRun.go
@@ -8,6 +8,7 @@ import (
 
 	"github.com/robocorp/rcc/cloud"
 	"github.com/robocorp/rcc/common"
+	"github.com/robocorp/rcc/conda"
 	"github.com/robocorp/rcc/operations"
 	"github.com/robocorp/rcc/pathlib"
 	"github.com/robocorp/rcc/pretty"
@@ -23,6 +24,7 @@ var assistantRunCmd = &cobra.Command{
 	Short:   "Robot Assistant run",
 	Long:    "Robot Assistant run.",
 	Run: func(cmd *cobra.Command, args []string) {
+		defer conda.RemoveCurrentTemp()
 		var status, reason string
 		status, reason = "ERROR", "UNKNOWN"
 		elapser := common.Stopwatch("Robot Assistant startup lasted")